[11/15/15] Code generated
Started generating code and writing implementation. Of course there are
still more components that need designed and generated, but I want to
get base functionality for communication finished so I can merge in this
branch.

Started with the AuthorizationManager to hammer out secure serialization
of internal objects. Will probably throw in some old code I have lying
around that did just that, but I want to do some research to make sure
that it's done the best way possible.